FIRST SECTOR IN MULTISECTOR READ IN BAD144 TABLE
	The bad144 table search code at finds the sector in the bad144
table and replaces the block number, cylinder, head, and sector addresses
with values corresponding to the replacement sector.  The sector count
register is loaded with the number of sectors in the entire transfer.
This is wrong; it *MUST* be set to *one* sector.  A read would return
the wrong data in sectors after the first; a write would *overwrite*
other replacement sectors or even the bad144 table on the last track.

AUTHOR:		Frank Maclachlan (fpm@crash.cts.com)
386BSD-Patchkit: patch00004