FIXES MANUAL PAGES FOR file AND magic COMMANDS
The manual page nroff sources for the file and magic command got munched
by mail or somesuch tool at one time.  All of the nroff commands had two
dots on them instead of one.  This patch corrects that.  The problem was
pointed out on the net and is in one of the buglist as bug MISC002.

AUTHOR:		Rodney W. Grimes (rgrimes@agora.rain.com)
386BSD-Patchkit: patch00103