Add the source code for fdisk, quotacheck, and mount_pcfs to sbin
	This patch adds the source code for the fdisk program that Julian
Elischer wrote.  Includes manual pages and lattest bug fixes as of 4/4/93.

	This patch adds the source code for the quotacheck program from
the Net/2 release tape.  There are 2 uninitialized variables in the Net/2
code.  dev_bsize is initialized in the wrong place; this bug only shows
itself if quotacheck is run on more than one disk partition.

	This patch adds the source code for the mount_pcfs program that
is part of the pcfs package by Pual Popelka.

AUTHOR: Julian Elischer (???) (fdisk)
AUTHOR:	Rodney W. Grimes (rgrimes@agora.rain.com) (Net/2 quotacheck)
AUTHOR: Paul Popelka (paulp@uts.amdahl.com) (mount_pcfs)
AUTHOR: Kent Talarico (kent@shipwreck.tsoft.net) (Net/2 quotacheck fixes)
386BSD-Patchkit: patch00111