Added the following to the  opening statement of install:
Before starting, it is important to know your hard disk's geometry
(i.e., number of cylinders, heads and sectors/track).  If installing
FreeBSD on the same disk as another operating system, then the
two systems should use the same geometry.  In particular, FreeBSD's
default geometry is inappropriate for MS-DOS.  So in this case, the
DOS geometry should be used instead.

[This seems to be true for SCSI disks.   What about IDE?  With the new
boot blocks, can we ignore the disks true geometry??]