Don't trust ncftp's exit status, they are not indicative of whether the file is fetched or not.  Apparently Jordan fixed it a long time ago but it was broken again at import of the new version of ncftp. Which means even if we fix it, it may break again and we may need to fix it again, and (imagination here, please)....
Instead, move the file existence check into the for loop for
MASTER_SITES/PATCH_SITES and break out with "continue 2" when the file
is found.  This is actually a cleaner logic than before if you ask me,
because instead of assuming the file is fetched on a 0 exit status
from ncftp AND checking for the existence of the file after the loop,
the check is done exactly once for each iteration and nowhere else.