Put back some missing chip identification code that got lost somehow during a trek through RCS. The Macronix 98713 and 98713A both have the same PCI device ID but different revision numbers, and we need to be able to tell one from the other. The 98715 and 98715A chips have the same device ID as the 98725 chip but different revision numbers, however we lump them into the same category except when identifying them during the PCI probe output.
The main reason we need tell the chips apart is that the Macronix app
notes say you have to write a special magic number into one of the
registers in order to put the chip in normal operating mode. The 98713
requires one magic value, while all the others require a different one.