MFC:
Make it possible to rebuild degraded RAID5 plexes.  Note that it is
currently not possible to do this while the volume is mounted.

Correctly skip the '/dev/' part when creating new drives and prefix
a drive's provider with '/dev/' when printing the config.

Don't allow to create a drive that already exists.

Approved by:  re@ (scottl)